move files around, couple small tweaks
