+++ app/src/react-app-env.d.ts
@@ -2,4 +2,4 @@
 declare module 'devextreme/data/query';
 declare module 'devextreme/localization';
 declare module '@apollo/client';
-declare module 'components/shared/api/Mutations';
+declare module 'api/Mutations';
